Split Unix module into per-command modules
